feat(db): added migration commands with alembic
