+++ pyproject.toml
@@ -27,6 +27,7 @@ classifiers = [
 ]
 requires-python = ">=3.12"
 dependencies = [
+    "alembic>=1.16.5",
     "apprise>=1.9.2",
     "copier>=9.7.1",
     "jinja2>=3.1.6",
