+++ src/kamihi/cli/commands/__init__.py
@@ -7,9 +7,10 @@ License:
 """
 
 from .action import app as action_app
+from .db import app as db_app
 from .init import app as init_app
 from .run import app as run_app
 from .user import app as user_app
 from .version import app as version_app
 
-__all__ = ["version_app", "action_app", "init_app", "run_app", "user_app"]
+__all__ = ["version_app", "action_app", "init_app", "run_app", "user_app", "db_app"]
