+++ src/kamihi/db/__init__.py
@@ -6,13 +6,14 @@ License:
 
 """
 
-from .db import init_engine, get_engine
+from .db import init_engine, get_engine, Base
 from .models import RegisteredAction, BaseUser, Role, Permission
 
 
 __all__ = [
     "init_engine",
     "get_engine",
+    "Base",
     "RegisteredAction",
     "BaseUser",
     "Role",
